requirements.txt
numpy
opencv-python
tensorflow
keras
pillow
tk
